refactor(front/components/Dashboard): new Dashboard for user and admin
