+++ client/src/components/Dashboard/DashboardAdmin.jsx
@@ -278,7 +278,7 @@ const useStyles = makeStyles((theme) => ({
   },
 }));
 
-export const Dashboard = () => {
+const Dashboard = () => {
   const dispatch = useDispatch();
   function createData(_id, name, price, stock) {
     return { _id, name, price, stock };
@@ -434,3 +434,5 @@ export const Dashboard = () => {
     </div>
   );
 };
+
+export default Dashboard;
